Extract course-listing into their own controller
