+++ app/views/events/courses/_nav_left.html.haml
@@ -3,5 +3,5 @@
     - @grouped_events.keys.each do |kind|
       %li= link_to(kind, "##{CGI.escape(kind)}", data: { turbolinks: false })
   - else
-    = render 'nav_left_events'
+    = render 'event/lists/nav_left_events'
