+++ config/routes.rb
@@ -1,5 +1,10 @@
 # frozen_string_literal: true
 
+#  Copyright (c) 2012-2021, Jungwacht Blauring Schweiz. This file is part of
+#  hitobito and licensed under the Affero General Public License version 3
+#  or later. See the COPYING file at the top-level directory or at
+#  https://github.com/hitobito/hitobito.
+
 Hitobito::Application.routes.draw do
 
   extend LanguageRouteScope
@@ -246,7 +251,7 @@ Hitobito::Application.routes.draw do
 
     end # resources :group
 
-    get 'list_courses' => 'event/lists#courses', as: :list_courses
+    get 'list_courses' => 'events/courses#index', as: :list_courses
     get 'list_events' => 'event/lists#events', as: :list_events
 
     get 'full' => 'full_text#index'
